feat: refactor backend to modular structure
