+++ backend/.gitignore
@@ -1,2 +1,3 @@
 .env
 __pycache__/
+venv/
